adds ToplineMetrics.svelte
