+++ src/components/explore/ProbeExplorer.svelte
@@ -9,20 +9,17 @@ import CompareOverTimeGraph from './CompareOverTimeGraph.svelte';
 import TotalClientsGraph from './TotalClientsGraph.svelte';
 import DistributionComparison from './DistributionComparison.svelte';
 import ComparisonSummary from './ComparisonSummary.svelte';
-import BigLabel from './BigLabel.svelte';
+import ToplineMetrics from './ToplineMetrics.svelte';
 
-
-import { toplineRefLabel, explorerComparisonSmallMultiple } from '../../utils/constants';
+import { explorerComparisonSmallMultiple } from '../../utils/constants';
 
 import {
-  formatBuildIDToDateString, formatCount, formatSignCount,
-  formatParenPercent,
+  formatBuildIDToDateString,
 } from '../../utils/formatters';
 
 import { histogramSpring } from '../../utils/animation';
 
 export let data;
-export let title;
 export let markers;
 export let key;
 export let timeHorizon;
@@ -95,10 +92,6 @@ $: if (densityMetricType && reference[densityMetricType]) {
   animatedReferenceDistribution.setValue(reference[densityMetricType]);
 }
 
-function absDiff(a, b, perc = false) {
-  return (a - b) / (perc ? b : 1);
-}
-
 
 function get(d, x) {
   return window1D({
@@ -158,57 +151,7 @@ h4 {
 
 
 <div class='probe-body-overview'>
-  <div style='
-      padding-left: {toplineRefLabel.left - toplineRefLabel.icon}px;
-      display: grid; grid-template-columns: max-content max-content max-content max-content; grid-column-gap: var(--space-2x);
-      font-size: var(--text-02);
-      align-items: start;
-      min-height: var(--space-8x);
-      align-content: start;
-    '>
-    <Tweenable params={{ duration: 250 }} value={reference.audienceSize} let:tweenValue>
-    <BigLabel value={reference.label}>
-      <span slot='icon'>⭑</span>
-      <span slot='label'>Reference</span>
-      <span slot='count'>
-          <span data-value={reference.audienceSize}>
-            <span style='font-weight: {hovered.datum && hovered.datum.audienceSize > tweenValue ? 'normal' : '500'}'>
-              {formatCount(tweenValue)}
-            </span> 
-            clients
-          </span>
-      </span>
-    </BigLabel>
-    <BigLabel params={{ duration: 0 }} 
-      value={hovered.datum ? hovered.datum.label : undefined}
-      compare={reference.label}
-    >
-      <span slot='icon'>●</span>
-      <span slot='label'>Hovered</span>
-      <span slot='count'>
-        <div>
-          {#if hovered.datum}
-          <span style='font-weight: {hovered.datum && hovered.datum.audienceSize < tweenValue ? 'normal' : '500'}'>
-            {formatCount(hovered.datum.audienceSize)}
-          </span> clients{/if}
-        </div>
-        {#if hovered.datum}
-        <div style="
-          font-weight:300;
-          white-space: pre;
-          color: {hovered.datum.audienceSize >= reference.audienceSize ? "var(--cool-gray-700)" : "var(--pantone-red-500)"}">
-{formatSignCount(absDiff(hovered.datum.audienceSize, tweenValue))}  <span style='font-weight: 500;'>{formatParenPercent(
-  '.0%',
-  absDiff(hovered.datum.audienceSize, tweenValue, true),
-  7,
-)}</span>
-        </div>
-        {/if}
-      </span>
-    </BigLabel>
-  </Tweenable>
-
-  </div>
+  <ToplineMetrics {reference} {hovered} />
   <slot name='summary'></slot>
 </div>
